==> 01_fbref_do_all_with_WC.ipynb <==
from fbref_get import *

players_22 = get_outfield_data(2022)
print('22 done')
players_21 = get_outfield_data(2021)
print('21 done')
players_20 = get_outfield_data(2020)
print('20 done')
players_19 = get_outfield_data(2019)
print('19 done')

wc_18 = pd.read_csv('fbref/wc_2018.csv', header=1)
wc_14 = pd.read_csv('fbref/wc_2014.csv', header=1)

fifa_data = pd.read_csv('fifa/fifa_22.csv')

players_22.columns

players_22 = players_22.dropna()
players_21 = players_21.dropna()
players_20 = players_20.dropna()
players_19 = players_19.dropna()

players_22['age'] = np.array([name[:-4] for name in players_22.age])

# players_22 = players_22[stats]
# players_21 = players_21[stats]
# players_20 = players_20[stats]
# players_19 = players_19[stats]

# wc_18 = wc_18[stats]
# wc_14 = wc_14[stats]

players_22['season'] = [2022] * len(players_22.player)
players_21['season'] = [2021] * len(players_21.player)
players_20['season'] = [2020] * len(players_20.player)
players_19['season'] = [2019] * len(players_19.player)

wc_18['season'] = [2018] * len(wc_18.Player)
wc_14['season'] = [2014] * len(wc_14.Player)

# players_22['player'] = np.array([name[:-9] for name in players_22.player])
# players_21['player'] = np.array([name[:-9] for name in players_21.player])
# players_20['player'] = np.array([name[:-9] for name in players_20.player])
# players_19['player'] = np.array([name[:-9] for name in players_19.player])

wc_18['Player'] = np.array([name[:-9] for name in wc_18.Player])
wc_14['Player'] = np.array([name[:-9] for name in wc_14.Player])

players = pd.concat([players_19, players_20, players_21, players_22], axis=0)

wc = pd.concat([wc_14, wc_18], axis=0)

player_names = np.unique(players.player)

sum_data = {}

for column_name in players.columns: 
    sum_data[column_name] = []

    
for player_name in player_names:

    df_player = players[players.player == player_name]

    for column_name in df_player.columns:
        
        if column_name == 'season': 
            sum_data[column_name].append(list(df_player.season)[-1])
        elif column_name == 'position':
            sum_data[column_name].append(list(df_player.position)[-1])
        elif column_name == 'squad':
            sum_data[column_name].append(list(df_player.squad)[-1])
        elif column_name == 'age':
            sum_data[column_name].append(int(list(df_player.age)[-1]))
        elif column_name == 'player':
            sum_data[column_name].append(list(df_player.player)[-1])
        elif column_name == 'nationality':
            sum_data[column_name].append(list(df_player.nationality)[-1])
        elif column_name == 'birth_year':
            sum_data[column_name].append(list(df_player.birth_year)[-1])
        else:
            sum_data[column_name].append(np.sum(df_player[column_name]))
            

df_sum = pd.DataFrame.from_dict(sum_data) # sum of domestic leagues

wc_data = {
    'wc_starts': [],
    'wc_games': [],
    'wc_minutes': [],
    'wc_goals': [],
    'wc_assists': [],
    'wc_cards_yellow': [],
    'wc_cards_red': [],
}

player_names = np.unique(df_sum.player)

non_wc_players = 0

for player_name in player_names:

    df_player = wc[wc.Player == player_name]
    
    if len(df_player.Player) == 0:
        non_wc_players += 1
        
        wc_data['wc_starts'].append(0)
        wc_data['wc_games'].append(0)
        wc_data['wc_minutes'].append(0)
        wc_data['wc_goals'].append(0)
        wc_data['wc_assists'].append(0)
        wc_data['wc_cards_yellow'].append(0)
        wc_data['wc_cards_red'].append(0)
    
    else:
        sum_stars = len(df_player.Player)
        sum_games = np.sum(df_player.Starts)
        sum_minutes = np.sum(df_player.Min)
        sum_goals = np.sum(df_player.Gls)
        sum_assists = np.sum(df_player.Ast)
        sum_yellow = np.sum(df_player.CrdY)
        sum_red = np.sum(df_player.CrdR)

        wc_data['wc_starts'].append(sum_stars)
        wc_data['wc_games'].append(sum_games)
        wc_data['wc_minutes'].append(sum_minutes)
        wc_data['wc_goals'].append(sum_goals)
        wc_data['wc_assists'].append(sum_assists)
        wc_data['wc_cards_yellow'].append(sum_yellow)
        wc_data['wc_cards_red'].append(sum_red)
        
print('Non WC players:', non_wc_players)
print('Total players:', len(wc_data['wc_starts']))

df_sum['wc_starts'] = wc_data['wc_starts']
df_sum['wc_games'] = wc_data['wc_games']
df_sum['wc_minutes'] = wc_data['wc_minutes']
df_sum['wc_goals'] = wc_data['wc_goals']
df_sum['wc_assists'] = wc_data['wc_assists']
df_sum['wc_cards_yellow'] = wc_data['wc_cards_yellow']
df_sum['wc_cards_red'] = wc_data['wc_cards_red']

team_names = [
    # La Liga
    ('Real Madrid','Real Madrid CF'), 
    ('Atlético Madrid','Atlético de Madrid'), 
    ('Real Sociedad','Real Sociedad'), 
    ('Sevilla','Sevilla FC'), 
    ('Betis','Real Betis Balompié'), 
    ('Rayo Vallecano','Rayo Vallecano'),
    ('Barcelona','FC Barcelona'),
    ('Athletic Club','Athletic Club de Bilbao'),
    ('Espanyol','RCD Espanyol de Barcelona'),
    ('Osasuna','CA Osasuna'),
    ('Valencia','Valencia CF'),
    ('Villarreal','Villarreal CF'),
    ('Celta Vigo','RC Celta de Vigo'),
    ('Mallorca','RCD Mallorca'),
    ('Alavés','Deportivo Alavés'),
    ('Granada','Granada CF'),
    ('Elche','Elche CF'),
    ('Cádiz','Cádiz CF'),
    ('Getafe','Getafe CF'),
    ('Levante','Levante Unión Deportiva'),
    # Bundesliga
    ('Bayern Munich','FC Bayern München'),
    ('Dortmund','Borussia Dortmund'),
    ('Leverkusen','Bayer 04 Leverkusen'),
    ('Hoffenheim','TSG 1899 Hoffenheim'),
    ('Freiburg','SC Freiburg'),
    ('Mainz 05','1. FSV Mainz 05'),
    ('Eint Frankfurt','Eintracht Frankfurt'),
    ('Union Berlin','1. FC Union Berlin'),
    ('RB Leipzig','RB Leipzig'),
    ('Köln','1. FC Köln'),
    ('Bochum','VfL Bochum 1848'),
    ('Wolfsburg','VfL Wolfsburg'),
    ('M\'Gladbach','Borussia Mönchengladbach'),
    ('Hertha BSC','Hertha BSC'),
    ('Stuttgart','VfB Stuttgart'),
    ('Augsburg','FC Augsburg'),
    ('Arminia','DSC Arminia Bielefeld'),
    ('Greuther Fürth','SpVgg Greuther Fürth'),
    # Premier League
    ('Manchester City','Manchester City'),
    ('Liverpool','Liverpool'),
    ('Chelsea','Chelsea'),
    ('Arsenal','Arsenal'),
    ('West Ham','West Ham United'),
    ('Manchester Utd','Manchester United'),
    ('Tottenham','Tottenham Hotspur'),
    ('Wolves','Wolverhampton Wanderers'),
    ('Leicester City','Leicester City'),
    ('Aston Villa','Aston Villa'),
    ('Crystal Palace','Crystal Palace'),
    ('Brentford','Brentford'),
    ('Brighton','Brighton & Hove Albion'),
    ('Everton','Everton'),
    ('Southampton','Southampton'),
    ('Leeds United','Leeds United'),
    ('Watford','Watford'),
    ('Burnley','Burnley'),
    ('Newcastle Utd','Newcastle United'),
    ('Norwich City','Norwich City'),
    # Serie A
    ('Inter','Inter'),
    ('Milan','Milan'),
    ('Atalanta','Atalanta'),
    ('Napoli','Napoli'),
    ('Fiorentina','Fiorentina'),
    ('Juventus','Juventus'),
    ('Lazio','Lazio'),
    ('Roma','Roma'),
    ('Empoli','Empoli'),
    ('Bologna','Bologna'),
    ('Sassuolo','Sassuolo'),
    ('Hellas Verona','Hellas Verona'),
    ('Torino','Torino'),
    ('Sampdoria','Sampdoria'),
    ('Udinese','Udinese'),
    ('Venezia','Venezia FC'),
    ('Spezia','Spezia'),
    ('Genoa','Genoa'),
    ('Cagliari','Cagliari'),
    ('Salernitana','US Salernitana 1919'),
    # Ligue 1
    ('Paris S-G','Paris Saint-Germain'),
    ('Marseille','Olympique de Marseille'),
    ('Rennes','Stade Rennais FC'),
    ('Nice','OGC Nice'),
    ('Montpellier','Montpellier HSC'),
    ('Lens','Racing Club de Lens'),
    ('Strasbourg','RC Strasbourg Alsace'),
    ('Monaco','AS Monaco'),
    ('Angers','Angers SCO'),
    ('Nantes','FC Nantes'),
    ('Lille','LOSC Lille'),
    ('Brest','Stade Brestois 29'),
    ('Lyon','Olympique Lyonnais'),
    ('Reims','Stade de Reims'),
    ('Bordeaux','FC Girondins de Bordeaux'),
    ('Clermont Foot','Clermont Foot 63'),
    ('Troyes','ESTAC Troyes'),
    ('Metz','FC Metz'),
    ('Lorient','FC Lorient'),
    ('Saint-Étienne','AS Saint-Étienne'),
]

for team in team_names:
    df_sum = df_sum.replace(team[0], team[1])

def get_fifa_data(df_fifa, df_team):

    def data_check(array, value):
        if len(np.array(value)) > 0:
            array.append(np.array(value)[0])
        else:
            array.append(np.nan)

        return array

    overall_values = []
    potential_values = []

    pace_values = []
    shooting_values = []
    passing_values = []
    dribbling_values = []
    defending_values = []
    physic_values = []
    
    wage_eur_values = []
    value_eur_values = []

    base = r'^{}'
    expr = '(?=.*{})'

    for name in np.array(df_team.player):
        names = name.split(' ')
        b = base.format(''.join(expr.format(w) for w in names))

        player = df_fifa[df_fifa.long_name.str.contains(b)]
        
        overall_values = data_check(overall_values, player.overall)
        potential_values = data_check(potential_values, player.potential)
                                      
        pace_values = data_check(pace_values, player.pace)
        shooting_values = data_check(shooting_values, player.shooting)
        passing_values = data_check(passing_values, player.passing)
        dribbling_values = data_check(dribbling_values, player.dribbling)
        defending_values = data_check(defending_values, player.defending)
        physic_values = data_check(physic_values, player.physic)
        
        wage_eur_values = data_check(wage_eur_values, player.wage_eur)
        value_eur_values = data_check(value_eur_values, player.value_eur)

    df_team["overall"] = overall_values
    df_team["potential"] = potential_values

    df_team["pace"] = pace_values
    df_team["shooting"] = shooting_values
    df_team["passing"] = passing_values
    df_team["dribbling"] = dribbling_values
    df_team["defending"] = defending_values
    df_team["physic"] = physic_values 
    
    df_team["wage_eur"] = wage_eur_values 
    df_team["value_eur"] = value_eur_values 

    return df_team

fifa_all = get_fifa_data(fifa_data, df_sum)

fifa_all.columns

fifa_all

fifa_all.to_csv('top5.csv')
