==> 03_stats_and_full_to_mongo.ipynb <==
import pandas as pd
import numpy as np

la_liga_teams = [
    ('Real Madrid','Real Madrid CF'), 
    ('Atlético Madrid','Atlético de Madrid'), 
    ('Real Sociedad','Real Sociedad'), 
    ('Sevilla','Sevilla FC'), 
    ('Betis','Real Betis Balompié'), 
    ('Rayo Vallecano','Rayo Vallecano'),
    ('Barcelona','FC Barcelona'),
    ('Athletic Club','Athletic Club de Bilbao'),
    ('Espanyol','RCD Espanyol de Barcelona'),
    ('Osasuna','CA Osasuna'),
    ('Valencia','Valencia CF'),
    ('Villarreal','Villarreal CF'),
    ('Celta Vigo','RC Celta de Vigo'),
    ('Mallorca','RCD Mallorca'),
    ('Alavés','Deportivo Alavés'),
    ('Granada','Granada CF'),
    ('Elche','Elche CF'),
    ('Cádiz','Cádiz CF'),
    ('Getafe','Getafe CF'),
    ('Levante','Levante Unión Deportiva'),
]

logos = [
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/53a2f082.png',
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/db3b9613.png',
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/e31d1cd9.png',
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/ad2be733.png',
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/fc536746.png',
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/98e8af82.png',
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/206d90db.png',
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/2b390eca.png',
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/a8661628.png',
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/03c57e2b.png',
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/dcc91a7b.png',
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/2a8183b3.png',
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/f25da7fb.png',
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/2aa12281.png',
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/8d6fd021.png',
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/a0435291.png',
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/6c8b07df.png',
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/ee7c297c.png',
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/7848bd64.png',
    'https://d2p3bygnnzw9w3.cloudfront.net/req/202201141/tlogo/fb/9800b6a1.png',
]

# Stat data processing

# read the data (statistics)
df_yigit = pd.read_csv("stat_data.csv", header=0)

# create a new header -- names for the columns

stats = ['min', '25', '50', '75', 'max']
features = ['goals', 'assists', 'gca', 'passes', 'aerials_won_pct', 'tackles']

header = ['squad', 'pos']

for feature in features:
    for stat in stats:
        c_name = feature + '_' + stat
        header.append(c_name)

# drop the remaining header rows in the dataframe

tmp = df_yigit.drop([0,1]) # drop the header rows

# create a dictionary for renaming the column names with something that makes sense

rename = {}
for old, new in zip(list(tmp.columns), header):
    rename[old] = new

tmp = tmp.rename(columns=rename)

# remove GKs from the dataframe
df_yigit = tmp[(tmp.pos == 'DF') | (tmp.pos == 'MF') | (tmp.pos == 'FW')]

df_yigit.head()

# FBref processing

df = pd.read_csv("top5_with_photos.csv")

df = df[list(df.columns)[1:]]

df = df[df.season == 2022]

df.head()

import json

file_name = 'laliga.json'

# makes JSON file for our MongoDB database
def team_dictionary(df_team, df_team_stats, team_name, logo):

    df_team = df_team.where(pd.notnull(df_team), None)
    df_team = df_team.replace({np.nan: None}) # <---
    players = df_team.T.to_dict()
    team = [players[key] for key in players]

    s = df_team_stats.T.to_dict()
    statistics = [s[key] for key in s]

    final_object = {
        'team': team_name,
        'logo': logo,
        'stats': statistics,
        'players': team
    }

    return final_object


league_dictionary = {
    'league_name': 'Spain Primera Division',
    'teams': [],
    }

counter = 0

for a, logo in zip(la_liga_teams, logos):

    _, team_name = a

    print(team_name)
    league_dictionary['teams'].append(
        team_dictionary(
            df[df.squad == team_name], 
            df_yigit[df_yigit.squad == team_name].drop('squad', axis=1).reset_index(drop=True),
            team_name,
            logo))
    
    counter += 1

with open(file_name, "w") as outfile:
    json.dump(league_dictionary, outfile)
